+++ meta/lint.el
@@ -2,7 +2,7 @@
 (find-file ".emacs.d/init.el")
 (next-line 2)
 (insert ";; Version: 0.1.0\n")
-(insert ";; Package-Requires: ((emacs \"29.1\"))\n")
+(insert ";; Package-Requires: ((emacs \"28.1\"))\n")
 (insert ";; URL: https://example.com/\n")
 (newline)
 (save-buffer)