+++ CHANGES.md
@@ -1,4 +1,4 @@
-0.4.0 (2024-08-29)
+0.4.0 (UNRELEASED)
 ------------------
 
 ### Changed
@@ -7,13 +7,15 @@
   loading until usage time.
 - Use `with-eval-after-load` to delay the evaluation of expressions
   that depend on certain packages to be loaded.
-- Use `setopt` instead of `setq`.
+- Replace `ido-mode` and `fido-mode` with `fido-vertical-mode`
+  (requires Emacs 28.1 or later).
 - Enable `show-trailing-whitespace` only in modes pertaining to
   configuration, programming, and text.
 - Introduce `install-packages` command to install packages.
 - Use `fboundp` to check if a mode exists before adding hooks that
   enable that mode.
 - Prevent Paredit from altering the behaviour of `RET`.
+
 
 
 0.3.0 (2023-06-22)